chore: configure v5
