+++ .babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": ["@babel/preset-env", "@babel/preset-react"],
+  "presets": ["@babel/preset-env", "@babel/preset-typescript", "@babel/preset-react"],
   "plugins": [
     "@babel/plugin-transform-object-assign",
     "@babel/plugin-proposal-object-rest-spread"
