+++ .eslintignore
@@ -1,5 +1,7 @@
 node_modules
 lib
 es
+lib-v5
+es-v5
 dist
 examples
