+++ index.js
@@ -3,3 +3,8 @@ module.exports.default = module.exports;
 module.exports.NextButton = require('./lib').NextButton;
 module.exports.PreviousButton = require('./lib').PreviousButton;
 module.exports.PagingDots = require('./lib').PagingDots;
+
+module.exports.v5 = require('./lib-v5').default
+module.exports.v5.NextButton = require('./lib-v5').NextButton;
+module.exports.v5.PreviousButton = require('./lib-v5').PreviousButton;
+module.exports.v5.PagingDots = require('./lib-v5').PagingDots;
