+++ package.json
@@ -21,6 +21,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.10.4",
     "@babel/plugin-transform-object-assign": "^7.10.4",
     "@babel/plugin-transform-runtime": "^7.10.4",
+    "@babel/plugin-transform-typescript": "^7.16.8",
     "@babel/polyfill": "^7.10.4",
     "@babel/preset-env": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
@@ -71,6 +72,10 @@
     "jsdom": "^14.0.0"
   },
   "scripts": {
+    "build-v5": "builder concurrent --buffer build-v5-lib build-v5-es",
+    "build-v5-babel": "babel src-v5 --extensions .ts,.tsx",
+    "build-v5-es": "builder run  --env \"{\\\"BABEL_ENV\\\":\\\"esm\\\"}\" build-v5-babel -- -d es-v5",
+    "build-v5-lib": "builder run build-v5-babel -- -d lib-v5",
     "build": "builder concurrent --buffer build-lib build-es",
     "build-babel": "babel src",
     "build-dist": "webpack --config webpack.dist.config.js --mode production",
