--- src-v5/all-transitions.ts
@@ -1,9 +1,0 @@
-import ScrollTransition from './transitions/scroll-transition';
-import FadeTransition from './transitions/fade-transition';
-import ScrollTransition3D from './transitions/3d-scroll-transition';
-
-export default {
-  fade: FadeTransition,
-  scroll: ScrollTransition,
-  scroll3d: ScrollTransition3D
-};
