+++ src-v5/default-carousel-props.tsx
